+++ src/Symfony/Bundle/TwigBundle/Tests/Functional/EmptyAppTest.php
@@ -69,7 +69,7 @@ class EmptyAppKernel extends Kernel
 
     public function getCacheDir(): string
     {
-        return sys_get_temp_dir().'/'.Kernel::VERSION.'/EmptyAppKernel/cache/'.$this->environment;
+        return sys_get_temp_dir().'/'.Kernel::VERSION.'/EmptyAppKernel/cache/'.$this->mode;
     }
 
     public function getLogDir(): string